inst/python/requirements.txt
# SAM3 Requirements
# Install SAM3 separately via: pip install git+https://github.com/facebookresearch/sam3.git

torch>=2.7
torchvision
rasterio>=1.3
geopandas>=0.14
shapely>=2.0
numpy>=1.24
pillow>=10.0
requests>=2.31
pyproj>=3.6
